bug fix of cocoex changes for Python 3
When moving the class definitions from interface to __init__
as inherited classes, a call of super was introduced. This
is neither necessary (__cinit__ is call implictely), nor
does it even work in Python 3, probably because no __init__
is implemented in the interface (only __cinit__). Removing
the super call resolves the problem.

	modified:   code-experiments/build/python/cython/interface.c
	modified:   code-experiments/build/python/python/__init__.py